+++ ring-multipart/src/deps.cljs
@@ -1,0 +1,1 @@
+{:npm-deps {"busboy" "^1.6.0"}}